finished dartboard kata
